+++ src/main/java/absolutelyaya/ultracraft/registry/PacketRegistry.java
@@ -50,10 +50,7 @@ import net.minecraft.text.Text;
 import net.minecraft.util.Hand;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Pair;
-import net.minecraft.util.hit.BlockHitResult;
-import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.math.*;
-import net.minecraft.world.RaycastContext;
 import net.minecraft.world.TeleportTarget;
 import net.minecraft.world.World;
 import org.apache.commons.lang3.ArrayUtils;
@@ -664,33 +661,21 @@ public class PacketRegistry
 					onTravelFinished(player);
 					return;
 				}
-				Pair<String, LevelManager.LevelInstance> instance = null;
+				Pair<String, LevelManager.LevelInstance> instance;
 				IWingedPlayerComponent winged = UltraComponents.WINGED.get(player);
-				if(instanceId.length() > 0)
+				if(instanceId.length() > 0 && LevelManager.Instance.isInstanceExistant(instanceId))
 				{
 					instance = new Pair<>(instanceId, LevelManager.Instance.getInstance(instanceId));
-					if(winged.getCurrentLevelInstance().equals(instanceId))
-					{
-						if(instance.getRight().getOwner().equals(player))
-							LevelManager.Instance.reloadInstance(instanceId);
-					}
-					LevelManager.Instance.onJoinLevel(player, instanceId);
+					if(winged.getCurrentLevelInstance().equals(instanceId) && instance.getRight().getOwner().equals(player))
+						LevelManager.Instance.reloadInstance(instanceId);
+					else
+						LevelManager.Instance.joinInstance(player, instanceId);
 				}
-				else if((instance = LevelManager.Instance.instantiateLevel(level)) == null)
-				{
+				else if((instance = LevelManager.Instance.instantiateLevel(level)) != null)
+					LevelManager.Instance.joinInstance(player, instance.getLeft());
+				else
 					player.sendMessage(Text.translatable("message.ultracraft.travel.error-instance"));
-					onTravelFinished(player);
-					return;
-				}
-				winged.enterLevel(level, instance.getLeft());
-				ServerWorld world = server.getWorld(LevelManager.WORLD_KEY);
-				BlockPos pos = LevelManager.getSpawnPos(instance.getLeft());
-				FabricDimensions.teleport(player, world, new TeleportTarget(pos.toCenterPos(), Vec3d.ZERO, LevelManager.getLevelData(level).getSpawnRot(), 0f));
-				BlockHitResult groundScan = player.getWorld().raycast(new RaycastContext(player.getPos(), player.getPos().subtract(0, 32, 0), RaycastContext.ShapeType.COLLIDER, RaycastContext.FluidHandling.NONE, player));
-				if(groundScan.getType().equals(HitResult.Type.MISS))
-					player.getWorld().setBlockState(player.getBlockPos().down(), BlockRegistry.PORTAL.getDefaultState());
 				onTravelFinished(player);
-				LevelManager.Instance.onJoinLevel(player, instance.getLeft());
 			});
 		});
 		ServerPlayNetworking.registerGlobalReceiver(PacketRegistry.REQUEST_DESTINATIONS_PACKET_ID, (server, player, handler, buf, sender) -> {